requirements.txt
opencv-python==4.8.0.74
ffmpeg-python==0.2.0
moviepy==1.0.3
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
tqdm==4.65.0

torch==2.0.1
torchvision==0.15.2
tensorflow==2.12.0

insightface==0.7.3
mediapipe==0.10.5
face-recognition==1.3.0
deepface==0.0.79
dlib==19.24.2

ultralytics==8.0.146
huggingface-hub==0.16.4
deepdanbooru==1.0.0

mmcv-full==1.7.1
mmpose==0.29.0

gradio==3.39.0
streamlit==1.25.0
matplotlib==3.7.2
plotly==5.15.0

git+https://github.com/openai/CLIP.git
# Note: Detectron2, OpenPose, OpenFace require manual installation